refactor(renderer): 抽 observation-inbox 11 个无闭包依赖 helper 到独立模块
把 observation-inbox-renderer.ts 顶层 2 个函数（feedbackAttributionRoleLabel、
renderFeedbackAttributionLabel）以及 renderObservationInboxPage 内部 9 个不依赖闭包
状态的纯 helper（skillAnchor、experienceSkillAnchor、renderJson、renderField、
formatTimestamp、truncateText、formatDuration、formatTimeRange、renderArtifactVersion）
集中迁移到 src/renderer/observation-inbox/helpers.ts。

formatDuration / formatTimeRange 需要的 lang 改为显式参数；renderer 内部通过一个
轻量 wrapper 闭包 formatTimeRange 预绑定 lang，保持函数体内所有调用点字节级不变。
renderFeedbackAttributionLabel 通过 renderer 顶部 re-export 保留对外 API。

HTML snapshot 字节一致；测量学红区文件零触碰。
